pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ApexVector"
version = "0.1.0"
description = "Python image vectorization: raster → optimized SVG via adaptive region classification"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Anshul Dhakar"}
]
keywords = ["vectorization", "svg", "image-processing", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "scikit-image>=0.19.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
apexvec = "apexvec.cli:main"

[project.urls]
Homepage = "https://github.com/Anshuldhakar1/ApexVector"
Repository = "https://github.com/Anshuldhakar1/ApexVector"
Issues = "https://github.com/Anshuldhakar1/ApexVector/issues"

[tool.setuptools.packages.find]
include = ["apexvec*"]
exclude = ["tests*", "vectorizer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
